Add auto primitiveShape to the webgl point feature.
The new default primitiveShape, `auto`, uses a sprite (a webgl point) if
points are smaller than the GPU's maximum point size and a triangle if
any are larger.  Specific primitive shapes can still be specified, as
there are performance reasons for each of them and auto detection has
some time cost.