Add a marker feature.
This adds an efficient webgl marker feature.  It uses slightly more
memory than the point feature (one more float value per marker is passed
to the GPU than for points).  A variety of shapes are defined: ellipses,
rectangles, isosceles triangles, and ovals, plus stellations of each of
these.  The shapes can specify orientation, and can automatically scale
and/or rotate with the map.

There is one distinct difference between markers and points.  The radius
of the marker is the total radius, including any stroke.  The size of
the point is the combination of the radius and the stroke.

These use nearly the same vertex shaders as the point feature, but a
vastly more complex fragment shader.  As such, rendering complex shapes
will saturate the GPU sooner than for basic points.

At some future point, it is intended to add image markers, where images
or svgs could be rendered via a texture map.  In this case, the intent
is to use a single texture buffer and pack multiple images into it with
a small space between each.

Resolves #820 (though a new issue should be created for image/svg
markers).