Add a strokeOffset option to markers.
This is useful, for instance, when using the ellipse marker and having
the ellipse scale with the map.  Setting the strokeOffset to 0 gives
behavior similar to a polygon.

Setting the strokeOffset to 1 allows setting the radius and strokeWidth
in the same manner as the pointFeature.

The default strokeOffset of -1 allows the radius to be the actual
maximal size of the marker.