Improve unstepped contours mapping of color and range.
For stepped data, the range values need to have one more entry than the
color values, since the range values define the divisions where colors
change.  For non-stepped data, we were still asking for one more range
entry than color entry, but the final range entry was not used.  Allow a
one-to-one mapping for non-stepped data, as this is more obvious.