feat: Add a grid feature.
This refactors the contourFeature to avoid code duplication, as the grid
and contour features are nearly the same.  Specifically, the contour
feature has data values at each vertex of a mesh and the grid feature
has data values on the elements of the mesh.  The grid feature has solid
colors for each mesh element, while the contour feature has interpolated
values between the vertices.

This adds a small tutorial.

This closes #545.