Refactor CI to not use cmake.
The principle benefit of using cmake was to download external test data.
This is simply handled by scripts/datastore.js.  All of the sha512 hash
references for test files have been moved there.

The secondary benefit from using cmake was integration with reporting to
cdash, but we haven't actually used cdash in any meaningful way.  We
don't use it to track status, so it doesn't provide any actual benefit.

The main motivator to moving away from cmake is to make it less complex
to shift CI from travis-ci to another service, such as circle-ci.