feat: Allow sharing queues between tile layers.
Although you could sort of share the fetch queue between tile layers, it
wouldn't really share properly without a custom needed function.  This
explicitly supports sharing a fetch queue between tile layers.

This allows, for instance, multiple layers that access the same server
to have a single fetch queue so that the number of requests against the
server is limited (usually to 6) rather than that number multiplied by
the number of layers.