feat: Improve when tile layers are idle.
Add a setting to allow tile layers to be considered idle either when all
tiles are finished loading or as soon as all tiles in the view are
loaded.  Before this was always equivalent to all tiles.  The new
default is just tiles in the view, as this is more relevant to typical
use.