perf: Speed up hit tests for points and markers
If the range tree returns multiple points or markers that might be
within the search area, we were sorting them to maintain a stable
response order.  Javascript sort defaults to a lexical sort which casts
all values to strings, so we had been doing `points.sort((a, b) => a
- b)` to do a numerical sort, but this has a lot of function call
overhead and probably is casting each value between strings and numeric
datatypes.  Instead, we do `points = Uint32Array.from(points).sort();`
which uses native javascript functions and is much faster.