fix: Fix pixelmap parameters for images smaller than tiles
Fix an issue when generating pixelmap parameters for images smaller than
tiles.  Specifically, if an image was small (say 100 x 100) and claimed
to be tiled where the tiles where more than twice the size of the image
(say 1024 x 1024), the minimum tile level was not set appropriately for
the tile layer, and, if corrected, the url could ask for a negative z
level.  If negative z levels are ever desired, this would need to
change.