fix: When lines reverse on themselves, mitering could be wrong
If a line reversed direction exactly, mitering logic could render only
half the pixels that should have been rendered.  In general, when lines
nearly reverse on themselves, mitering is tricky.  This add some
thresholds, below which the lines aren't mitered but are just rendered.
For opaque lines, this is a strict improvement, though some joints may
no longer show rounded or truncated miters.  For semi-transparent lines,
this is in general an improvement, but there can be cases where they
appear double rendered instead of joined.  The behavior is still deemed
an improvement, as they should never have only have half the pixels
rendered.