fix: Fix layer parameters for images smaller than tiles
In PR #1359, a problem was fixed when generating pixelmap parameters for
images smaller than tiles.  However, when a pixelmap is overlaid on a
lower resolution base, this changed the behavior of how that was
handled.  This fix specifically allows the lowest tile layer in a tile
layer as the minimum rather than zero.