fix: Change modifier keys for wheel actions
Since 2021-09 it has become the unofficial standard to send pinch events
on touchpads as wheel events with the ctrl key held down.  This should
correspond to zoom.  GeoJS actions predate this and had used ctrl+wheel
as rotate (and wheel by itself as zoom).  This changes it so shift+wheel
is rotate, and wheel+(no shift) regardless of ctrl state is zoom.