docs: Guard tutorial urls
When using a tutorial where we update the url to include the source, the
url can end in characters that some services misinterpret (e.g., if the
url ends in `.`, slack assumes that the `.` is not part of the url).  If
the last `src` chunk added does not end in an alphanumeric value, add a
`_=_` property to guard the end of the url.