perf: Add more information to annotation mode change events
Specifically, this add "oldCoordinates" to the event. If there was an
active annotation before the mode change, these are the annotation's
coordinates before the change.  This will be an empty list if an
annotation in create state had not been started and a non-empty list if
it is partially created.