fix: Fix rendering tiles that started off-map
When a tiles do not cover the entire map and start off screen, the
bookkeeping could load a tile and then the tile could be purged as
unneeded by still in the list of "don't check again" and not be loaded
when needed.  This currently uses an order n operation that could be
improved with a dictionary for performance if needed.